Get working regression, database manager and api test working
