+++ inputs/3lane.yml
@@ -3,11 +3,11 @@
 jobname: 3-lane
 type: "continuous" # differentiates between adding cars upfront or online
 time: 4000
-timestep: 0.1
+timestep: 0.5
 seed: 70
 logtype: "db"
 highway-type: "cpu" # "cpu", "kokkos" or "metal".
-road-end: 10000 # All lanes have a hard cutoff at 5000
+road-end: 10000 # All lanes have a hard cutoff at x = 10000
 
 # Driver Model
 driverType: "Gipps" # driver type
@@ -33,12 +33,12 @@ lanes:
 
     start: 0
     end: 5000
-    position: 1 # left lane
+    position: 0 # left lane
 
   - 
     flow:
       rate: 400 # flow rate (veh/hr)
-      v0: 30 # 0 represents all cars are starting from a standstill (stoplight case)
+      v0: 30
       vdes: 35 # all cars want to go 35 m/s
 
     start: 0
