+++ tests/CMakeLists.txt
@@ -16,6 +16,9 @@ find_package(libpqxx CONFIG REQUIRED)
 find_package(OpenSSL)
 find_package(CURL)
 
+add_library(testutil STATIC testUtil.cpp)
+target_link_libraries(testutil PUBLIC yaml-cpp::yaml-cpp PRIVATE pqxx_wrapper)
+
 add_executable(parsing parsing.cpp)
 target_link_libraries(parsing PUBLIC simulation 
                                      yaml-cpp::yaml-cpp 
@@ -24,15 +27,17 @@ target_link_libraries(parsing PUBLIC simulation
 
 add_executable(regression regression.cpp)
 target_compile_definitions(regression PUBLIC $<IF:$<BOOL:${OpenSSL_FOUND}>,"WITH_OPEN_SSL","">)
-target_link_libraries(regression PUBLIC simulation 
+target_link_libraries(regression PRIVATE simulation 
                                         jobManager
+                                        testutil
                                         $<IF:$<BOOL:OpenSSL_FOUND>,OpenSSL::Crypto,""> 
                                         GTest::gtest 
                                         GTest::gtest_main
                                         pqxx_wrapper)
 
 add_executable(dbManagerTest dbManager.cpp)
-target_link_libraries(dbManagerTest PUBLIC simulation 
+target_link_libraries(dbManagerTest PRIVATE simulation 
+                                      testutil
                                       jobManager 
                                       dbmanager
                                       GTest::gtest 
@@ -50,7 +55,7 @@ if (CURL_FOUND)
     add_executable(apiTest apiTest.cpp)
     target_link_libraries(apiTest PUBLIC api
                                         nlohmann_json::nlohmann_json
-                                        yaml-cpp::yaml-cpp
+                                        testutil
                                         GTest::gtest
                                         GTest::gtest_main
                                 PRIVATE CURL::libcurl
