+++ tests/dbManager.cpp
@@ -7,72 +7,19 @@
 #include "api/jobManager.hpp"
 #include "api/structs.hpp"
 #include "yaml-cpp/yaml.h"
+#include "testUtil.hpp"
 
 #include <pqxx/pqxx>
 
 class DBManagerTest : public ::testing::Test {
 
     protected:
-    YAML::Node getConfigNode() {
-        YAML::Node cfg;
-        cfg["type"] = "continuous";
-        cfg["time"] = 150;
-        cfg["timestep"] = 1;
-
-        // Driver params (From traffic flow book example)
-        cfg["driverType"] = "Gipps";
-        cfg["driverParams"]["a"] = 1.981;
-        cfg["driverParams"]["b"] = -2.8955;
-        cfg["driverParams"]["bmax"] = -5.505;
-        cfg["driverParams"]["p"] = 0.2;
-
-        // Homogeneous traffic
-        cfg["driverParams"]["a_stdev"] = 0;
-        cfg["driverParams"]["a_stdev"] = 0;
-        cfg["driverParams"]["bmax_stdev"] = 0;
-        cfg["driverParams"]["p_stdev"] = 0;
-
-
-        cfg["lanes"];
-        YAML::Node lane1, lane2;
-
-        lane1["flow"]["rate"] = 800;
-        lane1["flow"]["v0"] = 20;
-        lane1["flow"]["vdes"] = 35;
-        lane1["start"] = 0;
-        lane1["end"] = 2000;
-        lane1["posiiton"]  = 0;
-
-        lane2["flow"]["rate"] = 400;
-        lane2["flow"]["v0"] = 0;
-        lane2["flow"]["vdes"] = 38;
-        lane2["start"] = 0;
-        lane2["end"] = 2000;
-        lane2["posiiton"]  = 1;
-
-        cfg["lanes"].push_back(lane1);
-        cfg["lanes"].push_back(lane2);
-
-        cfg["logtype"] = "test";
-
-        return cfg;
-    }
-
-    void clearDB() {
-        // Clear out the Test DB:
-        pqxx::connection connect("host=localhost port=5432 dbname=trafficDBTest");
-        pqxx::work tx(connect);
-
-        tx.exec("DROP TABLE IF EXISTS trafficjobs CASCADE");
-        tx.exec("DROP TABLE IF EXISTS cardata CASCADE");
-        tx.exec("DROP TABLE IF EXISTS snapshotData");
-        tx.commit();
-    }
 
     void SetUp() override {
 
         for (size_t i = 0; i < 3; ++i){
-            YAML::Node dbLog = getConfigNode();
+            YAML::Node dbLog = TestUtil::getConfigNode();
+            dbLog["logtype"] = "test";
             dbLog["jobname"] = std::format("test-dbreader{}", i);
             dbLog["seed"] = 70 + i;
 
@@ -82,7 +29,7 @@ class DBManagerTest : public ::testing::Test {
             dbCfg << dbout.c_str();
         }
 
-        clearDB();
+        TestUtil::clearDB();
 
         // Run the tests with the job scheduler. 
         JobManager j;
@@ -99,16 +46,17 @@ class DBManagerTest : public ::testing::Test {
     void TearDown() override {
         for (size_t i = 0; i < 3; ++i){
             std::filesystem::path fname = std::format("dbConfig{}.yaml", i);
-            // if (std::filesystem::exists(fname)) std::filesystem::remove(fname);
+            if (std::filesystem::exists(fname)) std::filesystem::remove(fname);
         }
     }
 
 };
 
+// Disabled until error condition found. 
 class ErrorLogTest : public DBManagerTest {
 
     void SetUp() override {
-        YAML::Node dbLog = getConfigNode();
+        YAML::Node dbLog = TestUtil::getConfigNode();
         dbLog["jobname"] = "test-dbreader3";
         dbLog["seed"] = 70;
         dbLog["flow"]["rate"] = 900;
@@ -119,7 +67,7 @@ class ErrorLogTest : public DBManagerTest {
         dbout << dbLog;
         dbCfg << dbout.c_str();
 
-        clearDB();
+        TestUtil::clearDB();
 
     }
 
@@ -216,7 +164,7 @@ TEST_F(DBManagerTest, evaulateDB){
     EXPECT_TRUE(jobsAll.value().empty());
 }
 
-TEST_F(ErrorLogTest, errorLogging){
+TEST_F(ErrorLogTest, DISABLED_errorLogging){
     // Need to see if the error exists. 
 
     std::expected<void, std::string> result = Traffic::Simulate("dbConfig3.yaml");
